chagne to int

+++ src/sum_array_device.cu
@@ -3,19 +3,19 @@
 #include "sum_array_device.cuh"
 
 __global__
-void naiveSumArray(const float *input, float *output, int n) {
-    float partial_sum = 0.f;
+void naiveSumArray(const int *input, int *output, int n) {
+    int partial_sum = 0.f;
     for (int i = threadIdx.x; i < n; i += blockDim.x)
     {
-        float val = input[i];
+        int val = input[i];
         partial_sum += val;
     }
     atomicAdd(output, partial_sum);
 }
 
 void cudaSumArray(
-    const float *d_input,
-    float *d_output,
+    const int *d_input,
+    int *d_output,
     int n,
     SumArrayImplementation type)
 {
